nix/nix-builds: Make --fallback the default
We sometimes see the build servers mysteriously return 410 Gone for some store
paths. This could potentially be due to a known bug in nix related to the
interaction of the import-from-derivation feature and garbage collection. The
bug has the effect of things sometimes being garbage collected by mistake.

To make our deployment system more robust this patch adds the --fallback flag
(which falls back to local building) by default to nix builds. An option is
introduced to be able to turn this off for individual nix builds in case local
building hurts performance too much.

Note that this does not affect the deployment shell which does its own
invocation of nix-build and nix-shell.